move quicklink decorator to blockutils
